requirements.txt
tensorflow==2.18.0
numpy==1.26.4
pyttsx3==2.98
pillow==10.4.0
opencv-python==4.11.0.86
pandas==2.2.3
scikit-learn==1.6.1
scipy==1.13.1
matplotlib==3.10.0
seaborn==0.13.2
